@@ -1,11 +1,8 @@
-import React from "react";
-import ReactDOM from "react-dom";
+import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import "./index.css";
 
-ReactDOM.render(
-  <React.StrictMode>
-    <App />
-  </React.StrictMode>,
-  document.getElementById("root"),
-);
+const domNode = document.getElementById("root");
+const root = ReactDOM.createRoot(domNode as HTMLElement);
+
+root.render(<App />);
